+++ CHANGELOG.md
@@ -4,12 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [dev-2.0.0-0]
+## [dev-2.0.0-2]
+- [WIP] Quiz Frontend Plugin Frontend
+- [WIP] Create Riddler class to handle Quiz business logic
+- [IMPORTANT] Rename Quizsessions to singular and use it also has DomainObject for quiz session / Riddler
+- [FEATURE] Updated Model classes to new TCA
+
+## [dev-2.0.0-1]
 - [IMPORTANT] Rework all model to comply to the relational database model
 - [IMPORTANT] Rework TCA inclusion
-- [FEATURE] Change icons to WACON Logo
 - [FEATURE] Add Flexform for Simplequiz Plugin
 - [FEATURE] Rework ext_localconf and ext_tables.php
+
+## [dev-2.0.0-0]
+- [FEATURE] Change icons to WACON Logo
 
 ## [1.0.0] - 2024-03-31